Begun refactoring api
