+++ demo/src/app.js
@@ -1,0 +1,7 @@
+import ReadSmore from '../../src'
+
+const readMores = document.querySelectorAll('.js-read-more')
+
+const rm = ReadSmore(readMores)
+
+rm.init()
