+++ demo/src/app.scss
@@ -1,0 +1,75 @@
+*, body, html {
+  margin: 0;
+  padding:0;
+  line-height: 1.5;
+}
+
+html, body {
+  background-color: #eee;
+  font-family: "cardo";
+}
+
+h2, h3 {
+  font-family: "Roboto";
+  font-size: 1.75em;
+  margin-bottom: 1.25em;
+}
+p {
+  font-size: 1.1em;
+  margin-bottom: 1.2em;
+  color: #3e3f42;
+  line-height: 1.6;
+}
+
+
+.page {
+  position: relative;
+  max-width: 40em;
+  margin: 2em auto;
+  background-color: #fff;
+  box-shadow: 0px 1px 1px 1px rgba(0,0,0,0.1);
+  
+  &__main {
+    padding: 3em;
+    
+    @media (min-width: 32em) {
+      padding: 6em 4em;
+    }
+  }
+  
+  &__pretitle {
+     display:block;
+     margin: 0 0 2em;
+  }
+  
+  &__pretitle-link{
+     color: #2d2dc7;
+     font-family: "Roboto";
+     text-decoration:none;
+     transition: color 0.4s ease;
+    
+    &:hover {
+      color: blue;
+    }
+  }
+}
+
+
+.read-more {
+  
+  &__link {
+    font-family: "Roboto";
+    font-weight: 700;
+  }
+}
+
+// Add is-line to wrapper to support a read more inline with the content.
+.read-more.is-inline,
+.read-more.is-inline p,
+.read-more.is-inline + span  {
+  display: inline;
+}
+
+.read-more.is-inline + span {
+  margin-left: 0.25em
+}
