+++ demo/src/index.html
@@ -1,0 +1,36 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8"/>
+    <title>ReadSmore.js</title>
+    <link rel="stylesheet" href="app.scss" />
+    <script type="module" src="app.js"></script>
+  </head>
+  <body>
+    <main>
+      <article class="page">
+        
+      <section class="page__main">
+       
+        <h2 class="page__title">A Vanilla JS Read More</h2>
+        
+        <div class="page__pretitle">
+          <a class="page__pretitle-link" href="http://stephenscaff.com/articles/2018/12/readmore-js/" target="_blank">More Info</a>
+          |
+    <a class="page__pretitle-link" href="https://github.com/stephenscaff/readmore" target="_blank">Repo</a>
+        </div>
+        
+        <div class="readmore js-read-more" data-rm-words="70">
+      <p>We choose to go to the moon. We choose to go to the moon in this decade and do the other things, not because they are easy, but because they are hard.</p> 
+      
+    <p>Because that goal will serve to organize and measure the best of our energies and skills, because that challenge is one that we are willing to accept, one we are unwilling to postpone, and one which we intend to win, and the others, too.</p>
+    
+    <p>It is for these reasons that I regard the decision last year to shift our efforts in space from low to high gear as among the most important decisions that will be made during my incumbency in the office of the Presidency.</p>
+    
+    <p>In the last 24 hours we have seen facilities now being created for the greatest and most complex exploration in man's history. </p>
+        </div>
+      </section>
+    </article>
+    </main>
+  </body>
+</html>
