+++ src/index.js
@@ -1,0 +1,2 @@
+import ReadSmore from './ReadSmore.js'
+export default ReadSmore
