+++ src/lib/index.js
@@ -1,0 +1,2 @@
+import ReadMore from './ReadSmore.js'
+export default ReadMore
